Tania merge ke main

+++ app/Http/Controllers/ImmunizationController.php
@@ -4,35 +4,41 @@ namespace App\Http\Controllers;
 
 use App\Models\Immunization;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class ImmunizationController extends Controller
 {
-    // public function __construct()
-    // {
-    //     // Hanya admin yang bisa mengakses controller ini
-    //     $this->middleware(function ($request, $next) {
-    //         if (Auth::check() && Auth::user()->role !== 'admin') {
-    //             // Jika role bukan admin, tampilkan 403 forbidden
-    //             abort(403, 'Anda tidak memiliki akses ke halaman ini.');
-    //         }
-    //         return $next($request);
-    //     });
-    // }
-
-    public function index()
+    public function index(Request $request)
     {
-        $immunizations = Immunization::all();
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
+        $query = Immunization::query();
+
+        if ($request->has('name') && $request->name != '') {
+            $query->where('name', 'like', '%' . $request->name . '%');
+        }
+
+        $immunizations = $query->get();
+
         return view('admin.immunizations.index', compact('immunizations'));
     }
 
     public function create()
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
         return view('admin.immunizations.create');
     }
 
     public function store(Request $request)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255',
             'age' => 'nullable|string|max:100',
@@ -50,14 +56,23 @@ class ImmunizationController extends Controller
 
     public function edit(Immunization $immunization)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
         if (!$immunization) {
             return redirect()->route('admin.immunizations.index')->with('error', 'Imunisasi tidak ditemukan.');
         }
+
         return view('admin.immunizations.edit', compact('immunization'));
     }
 
     public function update(Request $request, Immunization $immunization)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255',
             'age' => 'nullable|string|max:100',
@@ -75,6 +90,10 @@ class ImmunizationController extends Controller
 
     public function destroy(Immunization $immunization)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized');
+        }
+
         $immunization->delete();
         return back()->with('success', 'Imunisasi berhasil dihapus');
     }
